Correct the link -- incorrectly.
I'm sure this isn't the right way to do it, but the resulting syntax
needs to be "projects/<project>/issues". I have managed to mimick this
by passing:

 {:controller => "projects", :action => project.identifier, :id =>
 "issues"}

However, this seems extremely hacky. However, in contrast, the menu
lines (in lib/redmine.rb) seem to stipulate that the following is
correct:

 {:controller => "issues", :action => "index"}

Using this instead resulted in a variety of links, none of which worked.
Thus it would seem to me that the menu itself is correct, but only
because it is called from already inside the projects controller.

One of these days it might help if I learnt Ruby and Rails.